rewrote extend routine
rewrote extend routine to have more descriptive variables, documentation, and more logical way of filling extension border. It also has an input parameter that can allow the definition of the number of additional control points by which to extend the array (previous was hard coded at 3). It defines the default behavior as to fill the extension border with the same displacements as the nearby edge, but can also be modified (setting extend_with_offsets to False) to fill the border with zero displacement.